% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMDB.R
\name{.read_hmdb_xml}
\alias{.read_hmdb_xml}
\title{.read_hmdb_xml}
\usage{
.read_hmdb_xml(xml_path, nonStop = TRUE)
}
\arguments{
\item{xml_path}{One xml file path.}

\item{nonStop}{Whether to stop when an error occurs.}
}
\value{
A data.frame.
}
\description{
.read_hmdb_xml
}
\examples{
xml_path <- system.file("demo", "HMDB0000014_ms_ms_spectrum_2248108_experimental.xml", package = "pubmsR")
.read_hmdb_xml(xml_path)
}
\author{
Barry Song
}
