% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_MsTb}
\alias{download_MsTb}
\title{Download MsTb from pubmsR}
\usage{
download_MsTb()
}
\description{
MsTb is stored with Git LFS, so user should use this function to download.
}
\examples{
download_MsTb()
}
