% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMDB.R
\name{load_hmdbMsTb}
\alias{load_hmdbMsTb}
\title{load_hmdbMsTb}
\usage{
load_hmdbMsTb(standard = TRUE)
}
\arguments{
\item{standard}{If TRUE, it will return a tibble with standard column names, which include accession, inchikey,
formula, name, exactmass, smiles, inchi, precursorMz, adduct, polarity, collision_energy, instrument_type, instrument, predicted,
mz, intensity.}
}
\value{
A hmdbMsTb tibble.
}
\description{
Load Human Metabolome Spectra Database (Version: 5.0)
The data is stored in a tibble with 65029 experimental and 65535 predicted spectra (18273 unique molecules),
which was generated form the xml file in \url{https://hmdb.ca/downloads}.
}
\examples{
hmdbMsTb <- load_hmdbMsTb()
}
\author{
Barry Song
}
