% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidBlast.R
\name{load_lipidblastMsTb}
\alias{load_lipidblastMsTb}
\title{Load LipidBlast MS/MS Tibble}
\usage{
load_lipidblastMsTb(standard = TRUE)
}
\arguments{
\item{standard}{If TRUE, it will return a tibble with standard column names, which include accession, inchikey,
formula, name, exactmass, smiles, inchi, precursorMz, adduct, polarity, collision_energy, instrument_type, instrument, predicted,
mz, intensity.}
}
\value{
A tibble.
}
\description{
Load LipidBlast MS/MS Spectra.
}
\examples{
lipidblastMsTb <- load_lipidblastMsTb()
}
