% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoNA.R
\name{load_monaMsTb}
\alias{load_monaMsTb}
\title{Load MoNA MS/MS Tibble}
\usage{
load_monaMsTb(standard = TRUE)
}
\arguments{
\item{standard}{If TRUE, it will return a tibble with standard column names, which include accession, inchikey,
formula, name, exactmass, smiles, inchi, precursorMz, adduct, polarity, collision_energy, instrument_type, instrument, predicted,
mz, intensity.}
}
\value{
A tibble.
}
\description{
Load MassBank of North America.
}
\examples{
monaMsTb <- load_monaMsTb()
}
