% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msp.R
\name{read_msp}
\alias{read_msp}
\title{Reading MSP files}
\usage{
read_msp(
  file_path,
  format = c("MoNA", "NIST", "RIKEN", "LipidBlast", "CUSTOM")[1],
  mapping = NA,
  thread = 1
)
}
\arguments{
\item{file_path}{msp file path.}

\item{format}{MoNA, NIST, RIKE and CUSTOM. When using CUSTOM, you should input your own mapping.}

\item{mapping}{If mapping is NA and format is CUSTOM, it will use MsBackendMsp default msp style.}

\item{thread}{Thread number.}
}
\value{
A data.frame.
}
\description{
Read msp file. There are several styles of msp format, they are MoNA, NIST and RIKEN etc.
}
\examples{
demo_MoNA_path <- system.file("demo", "demo_MoNA.msp", package = "pubmsR")
demo_NIST_path <- system.file("demo", "demo_NIST.msp", package = "pubmsR")
demo_RIKEN_path <- system.file("demo", "demo_RIKEN.msp", package = "pubmsR")
demo_LipidBlast_path <- system.file("demo", "demo_LipidBlast.msp", package = "pubmsR")
read_msp(demo_MoNA_path, format = "MoNA", thread = 1)
read_msp(demo_NIST_path, format = "NIST", thread = 1)
read_msp(demo_RIKEN_path, format = "RIKEN", thread = 1)
read_msp(demo_LipidBlast_path, format = "LipidBlast", thread = 1)
}
